Move Add Mountains in ascent modal to embeded modal
